01152022: Add routers and group swagger endpoints by type
